Scala 3: '(_ :Int) max 1' seems to be inferred to be BigInt...

+++ src/main/scala/com/argondesign/alogic/ast/ExprOps.scala
@@ -55,7 +55,7 @@ trait ExprOps { this: Expr =>
   }
 
   final private def mkIndex(idx: Int)(implicit cc: CompilerContext) = {
-    fix(ExprInt(false, clog2(tpe.shapeIter.next) max 1, idx))
+    fix(ExprInt(false, (clog2(tpe.shapeIter.next) max 1).toInt, idx))
   }
 
   final private def mkSliceLength(idx: Int)(implicit cc: CompilerContext) = {
